pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentlens"
version = "0.1.0"
description = "AI Agent Evaluation & Observability Platform"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-experimental>=0.3.0",
    "duckduckgo-search>=6.0.0",
    "opentelemetry-api==1.32.0",
    "opentelemetry-sdk==1.32.0",
    "opentelemetry-exporter-otlp==1.32.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentlens"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
